# UMAP with new Spectral Embedding initialization