--- a/cpp/include/cuml/manifold/spectral_embedding.hpp
+++ b/cpp/include/cuml/manifold/spectral_embedding.hpp
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+#include <raft/core/device_coo_matrix.hpp>
 #include <raft/core/device_mdspan.hpp>
 #include <raft/core/resources.hpp>
 
@@ -51,5 +52,10 @@ void transform(raft::resources const& handle,
                ML::SpectralEmbedding::params config,
                raft::device_matrix_view<float, int, raft::row_major> dataset,
                raft::device_matrix_view<float, int, raft::col_major> embedding);
+
+void transform_precomputed(raft::resources const& handle,
+                           ML::SpectralEmbedding::params config,
+                           raft::device_coo_matrix<float, int, int, int>& connectivity_graph,
+                           raft::device_matrix_view<float, int, raft::col_major> embedding);
 
 }  // namespace ML::SpectralEmbedding