--- a/cpp/include/cuml/manifold/spectral_embedding.hpp
+++ b/cpp/include/cuml/manifold/spectral_embedding.hpp
@@ -53,9 +53,9 @@ void transform(raft::resources const& handle,
                raft::device_matrix_view<float, int, raft::row_major> dataset,
                raft::device_matrix_view<float, int, raft::col_major> embedding);
 
-void transform_precomputed(raft::resources const& handle,
-                           ML::SpectralEmbedding::params config,
-                           raft::device_coo_matrix<float, int, int, int>& connectivity_graph,
-                           raft::device_matrix_view<float, int, raft::col_major> embedding);
+void transform(raft::resources const& handle,
+               ML::SpectralEmbedding::params config,
+               raft::device_coo_matrix<float, int, int, int>& connectivity_graph,
+               raft::device_matrix_view<float, int, raft::col_major> embedding);
 
 }  // namespace ML::SpectralEmbedding