--- a/cpp/src/umap/init_embed/spectral_algo.cuh
+++ b/cpp/src/umap/init_embed/spectral_algo.cuh
@@ -59,12 +59,10 @@ void launcher(const raft::handle_t& handle,
   ASSERT(n > static_cast<nnz_t>(params->n_components),
          "Spectral layout requires n_samples > n_components");
 
-  auto connectivity_graph =
-    raft::make_device_coo_matrix<float, int, int, int>(handle, n, n, coo->nnz);
-
-  raft::copy(connectivity_graph.structure_view().get_rows().data(), coo->rows(), coo->nnz, stream);
-  raft::copy(connectivity_graph.structure_view().get_cols().data(), coo->cols(), coo->nnz, stream);
-  raft::copy(connectivity_graph.get_elements().data(), coo->vals(), coo->nnz, stream);
+  auto connectivity_graph_view = raft::make_device_coo_matrix_view<float, int, int, int>(
+    coo->vals(),
+    raft::make_device_coordinate_structure_view<int, int, int>(
+      coo->rows(), coo->cols(), n, n, coo->nnz));
 
   ML::SpectralEmbedding::params spectral_params;
   spectral_params.n_neighbors    = params->n_neighbors;
@@ -78,7 +76,8 @@ void launcher(const raft::handle_t& handle,
   auto tmp_embedding_view = raft::make_device_matrix_view<float, int, raft::col_major>(
     tmp_embedding.data_handle(), n, params->n_components);
 
-  ML::SpectralEmbedding::transform(handle, spectral_params, connectivity_graph, tmp_embedding_view);
+  ML::SpectralEmbedding::transform(
+    handle, spectral_params, connectivity_graph_view, tmp_embedding_view);
 
   raft::linalg::transpose(
     handle, tmp_embedding.data_handle(), embedding, n, params->n_components, stream);