--- a/cpp/src/umap/init_embed/spectral_algo.cuh
+++ b/cpp/src/umap/init_embed/spectral_algo.cuh
@@ -21,8 +21,14 @@
 
 #include <raft/core/device_coo_matrix.hpp>
 #include <raft/core/handle.hpp>
+#include <raft/linalg/add.cuh>
 #include <raft/linalg/transpose.cuh>
+#include <raft/random/rng.cuh>
 #include <raft/sparse/coo.hpp>
+
+#include <thrust/device_ptr.h>
+#include <thrust/execution_policy.h>
+#include <thrust/extrema.h>
 
 #include <stdint.h>
 
@@ -55,6 +61,13 @@ void launcher(const raft::handle_t& handle,
     raft::make_device_coordinate_structure_view<int, int, int>(
       coo->rows(), coo->cols(), n, n, coo->nnz));
 
+  // raft::print_device_vector("connectivity_graph.elements",
+  // connectivity_graph_view.get_elements().data(), coo->nnz, std::cout);
+  // raft::print_device_vector("connectivity_graph.rows",
+  // connectivity_graph_view.structure_view().get_rows().data(), coo->nnz, std::cout);
+  // raft::print_device_vector("connectivity_graph.cols",
+  // connectivity_graph_view.structure_view().get_cols().data(), coo->nnz, std::cout);
+
   ML::SpectralEmbedding::params spectral_params;
   spectral_params.n_neighbors    = params->n_neighbors;
   spectral_params.norm_laplacian = true;
@@ -72,6 +85,33 @@ void launcher(const raft::handle_t& handle,
 
   raft::linalg::transpose(
     handle, tmp_embedding.data_handle(), embedding, n, params->n_components, stream);
+
+  uint64_t seed    = params->random_state;
+  auto tmp_storage = tmp_embedding;
+  raft::linalg::unaryOp<float>(
+    tmp_storage.data_handle(),
+    tmp_storage.data_handle(),
+    n * params->n_components,
+    [=] __device__(float input) { return fabsf(input); },
+    stream);
+
+  thrust::device_ptr<float> d_ptr = thrust::device_pointer_cast(tmp_storage.data_handle());
+  float max =
+    *(thrust::max_element(thrust::cuda::par.on(stream), d_ptr, d_ptr + (n * params->n_components)));
+
+  // Reuse tmp_storage to add random noise
+  raft::random::Rng r(seed);
+  r.normal(tmp_storage.data_handle(), n * params->n_components, 0.0f, 0.0001f, stream);
+
+  raft::linalg::unaryOp<float>(
+    embedding,
+    embedding,
+    n * params->n_components,
+    [=] __device__(float input) { return (10.0f / max) * input; },
+    stream);
+
+  raft::linalg::add(
+    embedding, embedding, tmp_storage.data_handle(), n * params->n_components, stream);
 
   RAFT_CUDA_TRY(cudaPeekAtLastError());
 }