Resolves #7052, Depends on https://github.com/rapidsai/cuvs/pull/1197

Use the new Spectral Embedding algorithm from `cuvs::preprocessing::spectral_embedding` for spectral initialization in UMAP. 
The previous spectral initialization has had longstanding issues mentioned in issues such as #5782.

# Visualization Analysis
Code beyond the initialization step in umap is commented out to obtain the plots https://github.com/rapidsai/cuml/blob/branch-25.08/cpp/src/umap/runner.cuh#L245C3-L258C40

1 = old spectral initialization
2 = new spectral initialization
3 = cpu spectral initialization using SpectralEmbedding()
4 = cpu spectral initialization

<img width="4714" height="1866" alt="image" src="https://github.com/user-attachments/assets/ed879dc3-9728-44a9-bd0f-995ff25f7637" />

<img width="7672" height="1860" alt="image" src="https://github.com/user-attachments/assets/4761f1dc-caf8-4ec7-a0cc-4db84d1e4e26" />

# Benchmarking 1x L4

| Dataset                      | Samples | Features | New Time (s)    | New Trustworthiness | Old Time (s)   | Old Trustworthiness |
|-----------------------------|---------|----------|------------------|----------------------|----------------|----------------------|
| deep-image-96-angular       | 50,000  | 96       | 0.321±0.004      | 0.8945±0.0007        | 0.341±0.006     | 0.8961±0.0013        |
| fashion-mnist-784-euclidean| 50,000  | 784      | 1.247±0.002      | 0.9743±0.0003        | 1.307±0.006     | 0.9726±0.0022        |
| gist-960-euclidean          | 50,000  | 960      | 1.642±0.003      | 0.7704±0.0010        | 1.702±0.017     | 0.7718±0.0011        |
| glove-25-angular            | 50,000  | 25       | 0.314±0.002      | 0.8146±0.0008        | 0.320±0.002     | 0.8191±0.0027        |
| mnist-784-euclidean         | 50,000  | 784      | 1.247±0.004      | 0.9522±0.0011        | 1.246±0.010     | 0.9528±0.0013        |
| sift-128-euclidean          | 50,000  | 128      | 0.326±0.003      | 0.9191±0.0008        | 0.325±0.003     | 0.9180±0.0009        |


| Dataset                      | Samples | Features | New Time (s)    | New Trustworthiness | Old Time (s)     | Old Trustworthiness |
|-----------------------------|---------|----------|-----------------|---------------------|------------------|---------------------|
| deep-image-96-angular       | 250,000 | 96       | 5.495±0.012     | 0.8891±0.0013       | 5.619±0.071      | 0.8873±0.0021       |
| fashion-mnist-784-euclidean| 60,000  | 784      | 1.782±0.004     | 0.9732±0.0003       | 1.872±0.027      | 0.9724±0.0014       |
| gist-960-euclidean          | 250,000 | 960      | 40.270±0.176    | 0.7656±0.0021       | 40.468±0.141     | 0.7660±0.0015       |
| glove-25-angular            | 250,000 | 25       | 5.128±0.004     | 0.8180±0.0024       | 5.201±0.015      | 0.8194±0.0023       |
| mnist-784-euclidean         | 60,000  | 784      | 1.803±0.010     | 0.9480±0.0022       | 1.794±0.006      | 0.9513±0.0017       |
| sift-128-euclidean          | 250,000 | 128      | 5.726±0.085     | 0.9156±0.0005       | 5.617±0.008      | 0.9150±0.0006       |



The following python script is used to obtain the plots.

```python3
import numpy as np
from sklearn.datasets import fetch_openml
import matplotlib.pyplot as plt
from cuml.manifold import UMAP, SpectralEmbedding
from sklearn.manifold import SpectralEmbedding as skSpectralEmbedding

mnist = fetch_openml('mnist_784', version=1, as_frame=False)
X_mnist = mnist.data.astype('float32')
y_mnist = mnist.target.astype(int)

subset_size = 5000
X = X_mnist
y = y_mnist

# Compute spectral embedding
spectral_model = SpectralEmbedding(n_components=2, n_neighbors=15)
X_spectral = spectral_model.fit_transform(X)

# Compute UMAP embedding
umap_model = UMAP(n_neighbors=15, n_components=2, random_state=42, n_epochs=0, init="spectral")
X_umap = umap_model.fit_transform(X)

# Plot side-by-side
fig, axs = plt.subplots(1, 2, figsize=(18, 7))

scatter1 = axs[0].scatter(X_spectral[:, 0], X_spectral[:, 1], c=y, cmap='tab10', s=2)
axs[0].set_title('cuML Spectral Embedding of MNIST')
axs[0].set_xlabel('Spectral Component 1')
axs[0].set_ylabel('Spectral Component 2')

scatter2 = axs[1].scatter(X_umap[:, 0], X_umap[:, 1], c=y, cmap='tab10', s=2)
axs[1].set_title('cuML UMAP Projection of MNIST')
axs[1].set_xlabel('UMAP 1')
axs[1].set_ylabel('UMAP 2')

plt.tight_layout()
plt.show()

```