@cjnolet 2025-07-31
@aamijar in addition to the viz above, please also benchmark the old and new spectral initialization on multiple datasets or different shapes and sizes so that we can get an idea of the perf gap (if any) between them.

it’s very very important that we characterize the delta here very carefully.