@aamijar 2025-08-03
Thanks for the review @viclafargue! I have addressed your comments and added the bounding box initialization constraint.