@viclafargue 2025-08-05
I think that the number of epochs has to be set to 1 to get a good visualization of spectral initialization within UMAP. Setting it to 0 [makes UMAP use a default value](https://github.com/rapidsai/cuml/blob/ff1c1afc1e2d825b51b4d945c71ab6937d6b50d7/cpp/src/umap/simpl_set_embed/algo.cuh#L310-L316).