@aamijar 2025-08-05
I am not running the simpl_set part since I comment out the part I mentioned in the PR description.