@viclafargue 2025-08-05
> I am not running the simpl_set part since I comment out the part I mentioned in the PR description.

Oh, I see, then it should be good. But, why are there only 3 points visible in the visualization then?