@cjnolet 2025-08-06
@aamijar thanks for updating the visualizations in the description. I'm a little confused by the second set of visualizations, though. Can you verify the "updated" visualization is what we would expect from the cpu version as well? You should be able to verify this by setting the number of epochs to 0, or printing the points in the cpu version right after initialization.