@aamijar 2025-08-08
Hi @cjnolet, I have added the benchmark results the PR description. I have also added additional plots to show the CPU spectral initialization. The CPU one also has 3 points if you run sklearn SpectralEmbedding (I edited the code to do this) as the spectral initialization. If you run the usual CPU spectral initialization you will get a different plot.